requirements.txt
bcrypt==4.0.1
blinker==1.5
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==1.1.3.post0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.4
SQLAlchemy==1.4.42
Werkzeug==2.2.2
WTForms==3.0.1
